test(api): move OIM upgrade coverage to migration runner
